Tweak code to follow coding standard
