+++ src/main/java/duke/command/Command.java
@@ -1,8 +1,8 @@
-package Command;
+package duke.command;
 
-import Utility.TaskList;
-import Utility.Ui;
-import Utility.Storage;
+import duke.utility.Storage;
+import duke.utility.TaskList;
+import duke.utility.Ui;
 
 public abstract class Command {
     public abstract void execute(TaskList list, Ui ui, Storage s);
