+++ src/main/java/duke/command/ExitCommand.java
@@ -1,8 +1,8 @@
-package Command;
+package duke.command;
 
-import Utility.TaskList;
-import Utility.Ui;
-import Utility.Storage;
+import duke.utility.TaskList;
+import duke.utility.Storage;
+import duke.utility.Ui;
 
 public class ExitCommand extends Command {
     public void execute(TaskList list, Ui ui, Storage s) {
