+++ src/main/java/duke/command/ListCommand.java
@@ -1,8 +1,8 @@
-package Command;
+package duke.command;
 
-import Utility.TaskList;
-import Utility.Ui;
-import Utility.Storage;
+import duke.utility.Ui;
+import duke.utility.TaskList;
+import duke.utility.Storage;
 
 public class ListCommand extends Command {
     public void execute(TaskList list, Ui ui, Storage s) {
