+++ src/main/java/duke/command/MarkCommand.java
@@ -1,22 +1,23 @@
-package Command;
+package duke.command;
 
 import java.io.IOException;
-import Exceptions.DukeException;
+import duke.exceptions.DukeException;
 
-import Utility.TaskList;
-import Utility.Ui;
-import Utility.Storage;
+import duke.utility.Ui;
+import duke.utility.TaskList;
+import duke.utility.Storage;
 
-public class UnmarkCommand extends Command {
+public class MarkCommand extends Command {
     private int taskNumber;
 
-    public UnmarkCommand(int i) {
+    public MarkCommand(int i) {
         this.taskNumber = i;
     }
+
     public void execute(TaskList list, Ui ui, Storage s) {
-        ui.showUnmark();
+        ui.showMark();
         try {
-            list.undo(this.taskNumber);
+            list.markDone(taskNumber);
             s.storeTaskList(list.getList());
         } catch (IOException | DukeException e) {
             ui.showError(e.getMessage());
