+++ src/main/java/duke/task/Task.java
@@ -1,8 +1,9 @@
-package Task;
+package duke.task;
 
 public class Task {
     private String name;
     private boolean isDone;
+
     public Task(String s) {
         this.name = s;
     }
@@ -14,6 +15,7 @@ public class Task {
     public void setDone(Boolean b) {
         this.isDone = b;
     }
+
     @Override
     public String toString() {
         return "[" + this.getStatusIcon() + "] " + this.name;
