+++ src/main/java/duke/task/Todo.java
@@ -1,7 +1,6 @@
-package Task;
+package duke.task;
 
 public class Todo extends Task {
-
     public Todo(String name) {
         super(name);
     }
