+++ src/test/java/Utility/ParserTest.java
@@ -1,6 +1,6 @@
-package Utility;
+package duke.utility;
 
-import Exceptions.DukeException;
+import duke.exceptions.DukeException;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
